Adding .drone.yml file, adding Dockerfile, moving index.js to server directory, updating specs and files references
