+++ docker/Dockerfile
@@ -1,0 +1,8 @@
+FROM node:8
+
+WORKDIR /opt/app
+COPY package.json ./
+RUN npm install
+COPY . .
+EXPOSE 3000
+CMD ["npm", "start"]
